% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_timelines.R
\name{cp_last_n_month}
\alias{cp_last_n_month}
\title{Custom Period Calculation - Last N Months}
\usage{
cp_last_n_month(
  data,
  monthyear,
  n,
  decimals = 7,
  att_column = "ATTRIBUTE",
  att_average = c("PDOVAL", "STR", "ND_OOS", "WD_OOS", "OOS_STORNO", "MSVAL", "MSVOL"),
  att_sum = c("SALVAL", "SALVOL", "SALQTY", "PURVOL", "PURQTY", "STKVOL", "STKQTY"),
  att_end_month = c("WDPERC", "NDPERC", "SAHPERC", "STORENO")
)
}
\arguments{
\item{data}{Transformed data frame where months are in the columns.}

\item{monthyear}{Years when Last N Months need to be calculated.}

\item{n}{Number of months to use for calculation.}

\item{decimals}{Number of decimal places to retain.}

\item{att_column}{Column name in the data frame that contains attributes.}

\item{att_average}{Attribute levels that need to be calculated by taking average.}

\item{att_sum}{Attribute levels that need to be calculated by taking sum.}

\item{att_end_month}{Attribute levels that need to be calculated by taking end_month.}
}
\value{
A data frame with Last N Months calculation added.
}
\description{
Custom Period Calculation - Last N Months
}
