% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_timelines.R
\name{cp_quarter}
\alias{cp_quarter}
\title{Custom Period Calculation - Quarters}
\usage{
cp_quarter(
  data,
  years = c(21:23),
  decimals = 7,
  att_column = "ATTRIBUTE",
  att_average = c("PDOVAL", "STR", "ND_OOS", "WD_OOS", "OOS_STORNO", "MSVAL", "MSVOL"),
  att_sum = c("SALVAL", "SALVOL", "SALQTY", "PURVOL", "PURQTY", "STKVOL", "STKQTY"),
  att_end_month = c("WDPERC", "NDPERC", "SAHPERC", "STORENO")
)
}
\arguments{
\item{data}{Transformed data frame where months are in the columns.}

\item{years}{Years when quarters need to be calculated. Format should be 2 digits number.}

\item{decimals}{Number of decimal places to retain.}

\item{att_column}{Column name in the data frame that contains attributes.}

\item{att_average}{Attribute levels that need to be calculated by taking average.}

\item{att_sum}{Attribute levels that need to be calculated by taking sum.}

\item{att_end_month}{Attribute levels that need to be calculated by taking end_month.}
}
\value{
A data frame with quarters added.
}
\description{
Custom Period Calculation - Quarters
}
