% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_handling.R
\name{from_excel_sheets}
\alias{from_excel_sheets}
\title{Import from Successive Excel Sheets}
\usage{
from_excel_sheets(path)
}
\arguments{
\item{path}{File path with file's name and extension.}
}
\value{
A data frame
}
\description{
Import from Successive Excel Sheets
}
