% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_share.R
\name{market_share}
\alias{market_share}
\title{Market Share Calculation}
\usage{
market_share(
  data,
  attribute = "Attribute",
  market = "MARKET",
  product_level = "PROD_LEVEL",
  val = "Sales Offtake Value (mil)",
  vol = "Sales Offtake Volume (ton)",
  val_share = "Market Share (Value)",
  vol_share = "Market Share (Volume)",
  top_level = "1_CATEGORY_TOTAL",
  decimals = 7,
  progress_full = FALSE
)
}
\arguments{
\item{data}{Data frame of the specific structure.}

\item{attribute}{Name of the column that contains the attributes.}

\item{market}{Name of the column that contains the markets.}

\item{product_level}{Name of the column that contains the product levels.}

\item{val}{Name of the attribute that contains sales offtake value in million.}

\item{vol}{Name of the attribute that contains sales offtake volume in tons.}

\item{val_share}{Name of the attribute that contains market share in terms of value.}

\item{vol_share}{Name of the attribute that contains market share in terms of volume.}

\item{top_level}{Name of the product level to use as base.}

\item{decimals}{Number of digits to retain after decimals.}

\item{progress_full}{To show progress in detail.}
}
\value{
A data frame with market share calculated.
}
\description{
Market Share Calculation
}
