% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_share.R
\name{market_share_single}
\alias{market_share_single}
\title{Market Share Calculation - Single Fact}
\usage{
market_share_single(
  data,
  market = "MARKET",
  product_level = "PROD_LEVEL",
  base_level = "1_CATEGORY_TOTAL",
  fact = "ATTRIBUTE",
  fact_amount = "Sales Offtake Value (mil)",
  fact_share = "Market Share (Value)",
  decimals = 7,
  progress_full = FALSE
)
}
\arguments{
\item{data}{Name of the data frame object.}

\item{market}{Column name that contains the markets.}

\item{product_level}{Column name that contains the product levels.}

\item{base_level}{Product level to use as denominator when calculating market share.}

\item{fact}{Column name that contains the facts}

\item{fact_amount}{Level among fact levels that contains amount.}

\item{fact_share}{Level among fact levels that contains market share.}

\item{decimals}{Numeric value that specifies how many digits to retain after decimal.}

\item{progress_full}{Logical. TRUE to show full process as the program loops through market and product levels.}
}
\value{
A data frame
}
\description{
Market Share Calculation - Single Fact
}
