% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mtcars}
\alias{mtcars}
\title{The Popular mtcars Dataset}
\format{
\subsection{\code{mtcars}}{

A data frame with 32 rows and 11 columns:
\describe{
\item{mpg}{Miles/(US) gallon}
\item{cyl}{Number of cylinders}
\item{disp}{Displacement (cu.in.)}
...
}
}
}
\source{
Henderson and Velleman (1981), Building multiple regression models interactively. Biometrics, 37, 391–411.
}
\usage{
mtcars
}
\description{
This is the mtcars dataset. This is added just as a demo.
}
\keyword{datasets}
