% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_handling.R
\name{split_by_column_labels}
\alias{split_by_column_labels}
\title{split_by_column_labels}
\usage{
split_by_column_labels(data, split_col, delete = TRUE)
}
\arguments{
\item{data}{Name of the data frame.}

\item{split_col}{Name of the column which should be used to split.}

\item{delete}{Logical. TRUE to delete the column which is used to split, FALSE otherwise.}
}
\value{
A list with data frames split by the specified column's levels.
}
\description{
split_by_column_labels
}
