% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{time_start}
\alias{time_start}
\title{Record Start Time of a Program}
\usage{
time_start()
}
\value{
Several objects loaded to the global environment
\itemize{
\item begin_time - Starting time of the program
\item end_time - Ending time of the program
\item runtime - Numeric value containing elapsed seconds
}
}
\description{
\code{time_start()} starts running the stopwatch and \code{time_stop()} stops the stopwatch to record run time.
}
\examples{
time_start()
Sys.sleep(10)
time_stop()
}
