% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformation.R
\name{transform_date_rows_to_cols}
\alias{transform_date_rows_to_cols}
\title{Title}
\usage{
transform_date_rows_to_cols(
  data,
  decimals = 7,
  date_col = "YM",
  div_100 = c("WDPERC", "NDPERC", "SAHPERC", "ND_OOS", "WD_OOS", "MSVAL", "MSVOL",
    "FMCGND"),
  div_1000 = c("SALVAL", "SALVOL", "PURVOL", "STKVOL"),
  mult_1000 = c("PDOVAL"),
  dup_attribute = "WDPERC",
  dup_market = "MARKET",
  dup_market_level = "ALL BGD",
  dup_item_desc = "ITEMDESC"
)
}
\arguments{
\item{data}{A data frame. The first column should contain date in the format 202312, i.e., YEARMONTH}

\item{decimals}{Number of decimal places to retain.}

\item{date_col}{Column name that refer to the column YEARMONTH, e.g., 202312.}

\item{div_100}{Columns that should be divided by 100.}

\item{div_1000}{Columns that should be divided by 1000.}

\item{mult_1000}{Columns that should be multiplied by 1000.}

\item{dup_attribute}{Name of any attribute level in the dataset. Just for calculating duplicates.}

\item{dup_market}{Column name that contains market information.}

\item{dup_market_level}{Name of the largest market, usually ALL BGD.}

\item{dup_item_desc}{Column name that contains item description.}
}
\value{
A transformed data frame.
}
\description{
Title
}
