% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_share.R
\name{market_share}
\alias{market_share}
\title{Market Share Calculation}
\usage{
market_share(
  data,
  market = "MARKET",
  product_level = "PROD_LEVEL",
  base_level = "1_CATEGORY_TOTAL",
  attribute = "ATTRIBUTE",
  val = "Sales Offtake Value (mil)",
  vol = "Sales Offtake Volume (ton)",
  val_share = "Market Share (Value)",
  vol_share = "Market Share (Volume)",
  decimals = 7,
  progress_full = FALSE
)
}
\arguments{
\item{data}{Name of the data frame object.}

\item{market}{Column name that contains the markets.}

\item{product_level}{Column name that contains the product levels.}

\item{base_level}{Product level to use as denominator when calculating market share.}

\item{attribute}{Column name that contains the attribute.}

\item{val}{Level among attribute levels that contains sales value.}

\item{vol}{Level among attribute levels that contains sales volume.}

\item{val_share}{Level among attribute levels that contains sales volume market share.}

\item{vol_share}{Level among attribute levels that contains sales value market share.}

\item{decimals}{Numeric value that specifies how many digits to retain after decimal.}

\item{progress_full}{Logical. TRUE to show full process as the program loops through market and product levels.}
}
\value{
A data frame
}
\description{
Market Share Calculation
}
